Split folder browser view sections
