+++ src/gui_app/folder_browser/view.rs
@@ -1,19 +1,23 @@
 use radiant::{
     prelude as ui,
-    widgets::{ButtonMessage, WidgetStyle, WidgetTone},
+    widgets::{WidgetStyle, WidgetTone},
 };
 
 use crate::gui_app::metadata_tags::{MetadataTagCompletionOption, MetadataTagDisplayCategory};
 
 use super::tag_editor::{metadata_section, tag_field_content_width, tag_field_height};
 use super::{
-    CollectionHitMessage, CollectionHitTarget, FolderBrowserMessage, FolderBrowserState,
-    GuiMessage, SourceEntry, TREE_DEPTH_INDENT, TREE_ROW_HEIGHT, VisibleFolder,
-    collections::COLLECTION_ROW_HEIGHT,
-    plural,
+    FolderBrowserMessage, FolderBrowserState, GuiMessage, TREE_DEPTH_INDENT, TREE_ROW_HEIGHT,
+    VisibleFolder, plural,
     tree_hit_target::{FolderTreeHitMessage, FolderTreeHitTarget},
     tree_widgets::FolderDropClearTarget,
 };
+
+mod collections_section;
+mod source_section;
+
+use collections_section::collections_section;
+use source_section::source_selector;
 
 pub(in crate::gui_app) fn folder_browser_view(
     state: &FolderBrowserState,
@@ -67,100 +71,6 @@ pub(in crate::gui_app) fn folder_browser_view(
     .fill_height()
 }
 
-fn collections_section(state: &FolderBrowserState) -> ui::View<GuiMessage> {
-    let rows = state
-        .visible_collections()
-        .into_iter()
-        .map(|collection| collection_row(state, collection))
-        .collect::<Vec<_>>();
-    sidebar_panel(
-        ui::column([
-            ui::row([
-                ui::text("Collections").height(20.0).fill_width(),
-                ui::drag_handle_mapped(|message| {
-                    GuiMessage::FolderBrowser(FolderBrowserMessage::ResizeCollectionsPanel(message))
-                })
-                .key("collections-resize-handle")
-                .size(26.0, 18.0),
-            ])
-            .spacing(4.0)
-            .height(20.0)
-            .fill_width(),
-            ui::scroll(ui::column(rows).spacing(1.0).fill_width().height(
-                COLLECTION_ROW_HEIGHT * wavecrate::sample_sources::SampleCollection::COUNT as f32,
-            ))
-            .style(WidgetStyle {
-                tone: WidgetTone::Neutral,
-                prominence: ui::WidgetProminence::Subtle,
-            })
-            .fill_width()
-            .fill_height(),
-        ])
-        .spacing(4.0)
-        .fill_width()
-        .fill_height(),
-        state.collections_panel_height(),
-    )
-}
-
-fn collection_row(
-    state: &FolderBrowserState,
-    collection: super::SampleCollectionView,
-) -> ui::View<GuiMessage> {
-    let collection_id = collection.collection;
-    if let Some(rename) = state.collection_rename_view(collection_id) {
-        let caret = rename.draft.chars().count();
-        return ui::row([
-            ui::custom_widget(CollectionHitTarget::new(&collection), |_| None)
-                .key(format!(
-                    "collection-rename-swatch-{}",
-                    collection.collection.index()
-                ))
-                .width(34.0)
-                .height(COLLECTION_ROW_HEIGHT),
-            ui::text_input(rename.draft)
-                .selection(0, caret)
-                .message_event(|message| {
-                    GuiMessage::FolderBrowser(FolderBrowserMessage::RenameInput(message))
-                })
-                .id(rename.input_id)
-                .key(format!(
-                    "collection-rename-input-{}",
-                    collection.collection.index()
-                ))
-                .fill_width()
-                .height(COLLECTION_ROW_HEIGHT),
-        ])
-        .key(format!(
-            "collection-rename-row-{}",
-            collection.collection.index()
-        ))
-        .fill_width()
-        .height(COLLECTION_ROW_HEIGHT)
-        .spacing(2.0);
-    }
-    ui::custom_widget_mapped(
-        CollectionHitTarget::new(&collection),
-        move |message| match message {
-            CollectionHitMessage::Activate => {
-                GuiMessage::FolderBrowser(FolderBrowserMessage::ActivateCollection(collection_id))
-            }
-            CollectionHitMessage::Rename => {
-                GuiMessage::FolderBrowser(FolderBrowserMessage::RenameCollection(collection_id))
-            }
-            CollectionHitMessage::Drop => {
-                GuiMessage::FolderBrowser(FolderBrowserMessage::DropOnCollection(collection_id))
-            }
-            CollectionHitMessage::HoverDropTarget(position) => GuiMessage::FolderBrowser(
-                FolderBrowserMessage::HoverCollectionDropTarget(collection_id, position),
-            ),
-        },
-    )
-    .key(format!("collection-row-{}", collection.collection.index()))
-    .fill_width()
-    .height(COLLECTION_ROW_HEIGHT)
-}
-
 fn folder_tree_view(state: &FolderBrowserState) -> ui::View<GuiMessage> {
     ui::stack([
         ui::custom_widget_mapped(
@@ -181,70 +91,6 @@ fn folder_tree_view(state: &FolderBrowserState) -> ui::View<GuiMessage> {
         .spacing(1.0),
     ])
     .fill()
-}
-
-fn source_selector(state: &FolderBrowserState) -> ui::View<GuiMessage> {
-    ui::column([
-        ui::row([
-            ui::text("Sources").height(20.0).fill_width(),
-            ui::button("+")
-                .primary()
-                .message(GuiMessage::FolderBrowser(FolderBrowserMessage::AddSource))
-                .key("source-add-button")
-                .size(28.0, 22.0),
-        ])
-        .spacing(3.0)
-        .fill_width()
-        .height(24.0),
-        ui::column(
-            state
-                .sources
-                .iter()
-                .map(|source| source_row(state, source))
-                .collect::<Vec<_>>(),
-        )
-        .spacing(2.0)
-        .fill_width(),
-    ])
-    .spacing(3.0)
-    .fill_width()
-}
-
-fn source_row(state: &FolderBrowserState, source: &SourceEntry) -> ui::View<GuiMessage> {
-    let id = source.id.clone();
-    let row_key = source.id.clone();
-    let menu_id = source.id.clone();
-    let selected = state.selected_source == source.id;
-    let label = if source.loading_task.is_some() {
-        format!("{} (scanning)", source.label)
-    } else {
-        source.label.clone()
-    };
-    let mut row = ui::button(label)
-        .secondary_clicks()
-        .mapped(move |message| match message {
-            ButtonMessage::SecondaryActivate { position } => GuiMessage::FolderBrowser(
-                FolderBrowserMessage::OpenSourceContextMenu(menu_id.clone(), position),
-            ),
-            _ => GuiMessage::FolderBrowser(FolderBrowserMessage::SelectSource(id.clone())),
-        })
-        .key(format!("source-row-{row_key}"))
-        .fill_width()
-        .height(24.0);
-    if selected {
-        row = row.primary();
-    } else {
-        row = row.subtle();
-    }
-    row.style(if selected {
-        WidgetStyle {
-            tone: WidgetTone::Accent,
-            prominence: ui::WidgetProminence::Subtle,
-        }
-    } else {
-        WidgetStyle::default()
-    })
-    .fill_width()
 }
 
 fn folder_row(folder: VisibleFolder) -> ui::View<GuiMessage> {
